+++ steam/webauth.py
@@ -63,7 +63,7 @@ import requests
 
 from steam import webapi
 from steam.steamid import SteamID
-from steam.util.web import make_requests_session, generate_session_id
+from steam.utils.web import make_requests_session, generate_session_id
 from steam.core.crypto import rsa_publickey, pkcs1v15_encrypt
 
 if six.PY2: